Locked down the USB share section such that random users to the Unplug won't be able to access any configuration, status or reboot options.
Used paths relative to the webroot ensuring that it works when resourced by an
authenticated administrator in the first screen/page of the share menu.

Provided an option/means of unsetting a saved WiFi bridge previously set to be
the device's default route/connection.

Menu items in the first detection menu toggle open/close again, like they're
supposed to. Small edit to admin/js/app.js.

  modified:   root/scripts/wifi.sh
  modified:   www/admin/bridge.php
  modified:   www/admin/cgi-bin/config.cgi
  modified:   www/admin/detect.php
  modified:   www/admin/header.php
  modified:   www/admin/js/app.js
  modified:   www/admin/start.php
  modified:   www/css/style.css
  new file:   www/index.html.share
  modified:   www/share/cgi-bin/config.cgi
  new file:   www/share/footer.php
  new file:   www/share/header.php
  modified:   www/share/index.php
  new file:   www/share/index.php.share